Transition from pattern-based to track-based workflow

+++ engine/src/modules/processing_graph/model/node.cpp
@@ -19,8 +19,9 @@
 
 #include "node.h"
 #include "generated/lib/model/model.h"
+#include "modules/processors/db_meter.h"
 
-std::optional<std::shared_ptr<NodePort>> Node::getPortById(int32_t id) {
+std::optional<std::shared_ptr<NodePort>> Node::getPortById(int64_t id) {
   for (auto& port : *this->audioInputPorts()) {
     if (port->id() == id) {
       return port;
